fix(linter): keep xargs parity shape in facts
